@W-22821837@ Move buildReturnProductItems into return-utils
Relocate the helper out of return-items-modal/constants.js (which now
holds only intl message definitions) into the existing return-utils.js
alongside getReturnableItems. Update the modal import and migrate the
helper-only unit tests to return-utils.test.js so the modal test file
stays focused on UI behavior.
